// lib/api-client/rq-generated/requests/types.gen.ts
export type ProductDto = {
    id?: string;
    imageUrl?: string;
    file?: (Blob | File);
    name?: string;
    description?: string;
    price: number;
    stockQuantity?: number;
    categoryNames?: Array<(string)>;
    sizes?: Array<(string)>;
};

export type ProductVariantDto = {
    id?: string;
    imageUrl?: string;
    file?: (Blob | File);
    colorCode?: string;
    productId?: string;
    price: number;
    stockQuantity?: number;
    sizes?: Array<(string)>;
};